pyproject.toml
[tool.poetry]
name = "template"
version = "0.1.0"
description = ""
authors = ["gokamoda <gokamoda@gmail.com>"]
readme = "README.md"
packages = [{include = "template"}]

[tool.poetry.dependencies]
python = "^3.10.5"
requests = "^2.31.0"
hydra-core = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"